08_agent_knowledge_system/c08_agent_knowledge_system/requirements.txt
# 基础依赖
python>=3.8
dashscope>=1.14.0
requests>=2.31.0
python-dotenv>=1.0.0
retrying>=1.3.4

# 向量库+嵌入模型
chromadb>=0.4.24
sentence-transformers>=2.7.0
numpy>=1.26.4

# 文档解析
pdfplumber>=0.10.3
python-markdown>=3.5.1
pytesseract>=0.3.10  # OCR（可选）
Pillow>=10.2.0       # 图片处理

# 数据库
sqlite3>=2.6.0